+++ src/Tests/Dfe.Complete.Tests/Models/ConversionTaskListViewModelTests.cs
@@ -654,6 +654,43 @@ namespace Dfe.Complete.Tests.Models
         }
 
         [Theory]
+        [InlineData(null, null, null, null, null, null, null, null, TaskListStatus.NotStarted)]
+        [InlineData(false, false, false, false, false, false, false, false, TaskListStatus.InProgress)]
+        [InlineData(true, true, true, true, true, true, true, false, TaskListStatus.Completed)]
+        [InlineData(null, null, null, null, null, null, null, true, TaskListStatus.NotApplicable)]
+        [InlineData(true, false, false, false, false, false, false, false, TaskListStatus.InProgress)]
+        [InlineData(false, false, true, true, true, true, true, false, TaskListStatus.Completed)]
+        public void PrivateFinanceInitiativeTaskStatus_ShouldReturn_CorrectStatus(
+            bool? supplementaryFundingAgreementPfiClausesInserted,
+            bool? masterFundingAgreementPfiClausesInserted,
+            bool? received,
+            bool? cleared,
+            bool? draftSavedInTrustSharepointFolder,
+            bool? signedByAllStakeholders,
+            bool? finalVersionSavedInSchoolAndTrustSharepointFolder,
+            bool? notApplicable,
+            TaskListStatus expectedStatus)
+        {
+            var taskData = new ConversionTaskDataDto
+            {
+                Id = new TaskDataId(Guid.NewGuid()),
+                PrivateFinanceInitiativeNotApplicable = notApplicable,
+                PrivateFinanceInitiativeSupplementaryFundingAgreementPfiClausesInserted = supplementaryFundingAgreementPfiClausesInserted,
+                PrivateFinanceInitiativeMasterFundingAgreementPfiClausesInserted = masterFundingAgreementPfiClausesInserted,
+                PrivateFinanceInitiativeReceived = received,
+                PrivateFinanceInitiativeCleared = cleared,
+                PrivateFinanceInitiativeDraftSavedInTrustSharepointFolder = draftSavedInTrustSharepointFolder,
+                PrivateFinanceInitiativeSignedByAllStakeholders = signedByAllStakeholders,
+                PrivateFinanceInitiativeFinalVersionSavedInSharepointFolder = finalVersionSavedInSchoolAndTrustSharepointFolder
+            };
+
+            var project = new ProjectDto();
+            var result = ConversionTaskListViewModel.Create(taskData, project, null);
+
+            Assert.Equal(expectedStatus, result.PrivateFinanceInitiative);
+        }
+
+        [Theory]
         [InlineData(null, TaskListStatus.NotStarted)]
         [InlineData(false, TaskListStatus.NotStarted)]
         [InlineData(true, TaskListStatus.Completed)]